Use the LoggableGraphqlError just in the interceptor
We need a different type for the generic use case
